+++ tests/test_ignition.py
@@ -6,7 +6,6 @@ EXAMPLE_ENGINE_URL_0 = "mysql://root:{env_secret_password}@localhost/my_database
 
 
 class TestIgnition(unittest.TestCase):
-
     def test_get_raw(self):
         """
         raw is always available because that's the url passed to constructor.